rename `document` module to `svg`
